+++ tools/detectThirdParties/tool.js
@@ -1,22 +1,11 @@
 (() => {
   class DetectThirdPartiesTool extends LiChessTools.Tools.ToolBase {
 
-    styleBlockContains = (text) => {
-      const lt = this.lichessTools;
-      const styleTags = document.querySelectorAll('style');
-      for (const style of styleTags) {
-        if (style.textContent.includes(text)) {
-          return true;
-        }
-      }
-      return false;
-    };
-
     async start() {
       const lt = this.lichessTools;
       const $ = lt.$;
       this.options = {
-        isPrettier: !!lt.global.getComputedStyle(lt.global.document.documentElement).getPropertyValue('--main-shadow'),
+        isPrettier: !!$('html').css('--boardLightRGB'),
         lichessHelper: !!$('style').get().find((s)=>$(s).text().includes('LichessHelper'))
       };