+++ tools/gameListOptions/tool.js
@@ -34,6 +34,7 @@
         'abortedGamesLabel': 'Show aborted:',
         'colorGamesLabel': 'Color by players:',
         'analysedGamesLabel': 'Only analysed:',
+        'notAnalysedGamesLabel': 'Only not analysed:',
         'analysisLinkLabel': 'Click to analysis:',
         'titledOpponentsLabel': 'Only titled opponents:',
         'copyGamesButtonTitle': 'Download selected games as PGN',
@@ -67,6 +68,7 @@ Inaccuracies / mistakes / blunders
         'abortedGamesLabel': 'Arat\u0103 anulate:',
         'colorGamesLabel': 'Culoare dup\u0103 juc\u0103tori:',
         'analysedGamesLabel': 'Doar analizate:',
+        'notAnalysedGamesLabel': 'Doar neanalizate:',
         'titledOpponentsLabel': 'Doar adversari cu titlu:',
         'analysisLinkLabel': 'Click pentru analiz\u0103:',
         'copyGamesButtonTitle': 'Descarc\u0103 jocurile selectate ca PGN',
@@ -337,6 +339,16 @@ Inexactit\u0103\u0163i/gre\u015feli/gafe
       return result.join('\r\n');
     };
 
+    rightClickGame = ev=>{
+      const lt = this.lichessTools;
+      const $ = lt.$;
+       const href = $(ev.currentTarget).closest('article.game-row').attr('data-orig-href');
+       if (href) {
+         ev.preventDefault();
+         lt.global.open(href, '_blank');
+       }
+    };
+
     processListsDirect = ()=>{
       const lt = this.lichessTools;
       const $ = lt.$;
@@ -366,6 +378,20 @@ Inexactit\u0103\u0163i/gre\u015feli/gafe
         }
       }
       if (this.options.analysis) {
+        if (!$(container).find('#chkNoAnalysis').length) {
+          filters
+            .prepend($('<input type="checkbox" id="chkNoAnalysis"/>')
+              .on('change',ev=>{
+                const checked = !!ev.target.checked;
+                lt.storage.set('LiChessTools.gameListOptions.noAnalysis',checked);
+                container.toggleClass('lichessTools-gameListOptions-noAnalysis',checked);
+                $('body').trigger('scroll');
+              })     
+              .prop('checked', !!lt.storage.get('LiChessTools.gameListOptions.noAnalysis'))
+            )
+            .prepend($('<label for="chkNoAnalysis"></label>').text(trans.noarg('notAnalysedGamesLabel')));
+          filters.find('#chkNoAnalysis').trigger('change');
+        }
         if (!$(container).find('#chkAnalysis').length) {
           filters
             .prepend($('<input type="checkbox" id="chkAnalysis"/>')
@@ -458,13 +484,17 @@ Inexactit\u0103\u0163i/gre\u015feli/gafe
           const xref = m
             ? '/' + m.groups.gameId.slice(0,8) + (isBlack?'/black':'')
             : href;
-          elem.attr('href',xref);
+          elem
+            .on('contextmenu',this.rightClickGame)
+            .attr('href',xref);
           $(e).attr('data-orig-href',href);
         }  
         if (!isAnalysisLink && $(e).attr('data-orig-href')) {
           const elem = $(e).find('a.game-row__overlay');
           const href = $(e).attr('data-orig-href');
-          elem.attr('href',href);
+          elem
+            .off('contextmenu',this.rightClickGame)
+            .attr('href',href);
           $(e).removeAttr('data-orig-href');
         }